+++ CHANGELOG.md
@@ -12,6 +12,8 @@ For notes on migrating to 2.x / 0.200.x see [the upgrade guide](doc/upgrade-to-2
 
 ### :boom: Breaking Changes
 
+* feat(exporter-jaeger)!: remove deprecated `@opentelemetry/exporter-jaeger` package [#6900](https://github.com/open-telemetry/opentelemetry-js/issues/6900) @harshmathurx
+
 ### :rocket: Features
 
 ### :bug: Bug Fixes